http2: improve error handling
Error handling from the binding layer was not very sophisticated. This
changes things such that errors on generally synchronous APIs are
thrown, while only errors that happen in async operations end up
in error events. This still may need to evolve but this is closer
to what it should be. Also, as much as possible, hide the details
of NGHTTP2 specific errors using Node.js specific error codes and
messages.

There is still more to be done here.