http2: make ncopy signed
Currently there is the following compiler warning generated upon
building:

../src/node_http2_core-inl.h:146:17: warning: comparison of integers of
different signs: 'ssize_t'
      (aka 'long') and 'size_t' (aka 'unsigned long') [-Wsign-compare]
      if (ncopy > buf.len)
          ~~~~~ ^ ~~~~~~~

ncopy is set to len which is of type ssize_t which is the return type of
nghttp2_session_mem_send, and there are checks to verify that len is
greater than zero so the converstion to size_t from ssize_t is safe.
ncopy can also be assigned to buf.len but uv_buf_t.len is of type
size_t.